feat: lwm q2 tour refactor
