+++ .changeset/product-tour-q2-drawer.md
@@ -2,4 +2,4 @@
 "live-mobile": minor
 ---
 
-Add the image-based Q2 Wallet V4 Tour drawer (mobile-only): swipeable slides with theme-aware (light/dark) lossless WebP illustrations, a single full-width footer button ("See what's new" / "Next" / "Got it"), gated by the `lwmWallet40` `q2Tour` parameter and persisted via `hasSeenQ2WalletV4Tour`. Wired into the "Debug Wallet v4" devtool for testing.
+Add the Q2 Wallet V4 Tour image carousel drawer on mobile, gated by `lwmWallet40.params.q2Tour`.
