+++ apps/ledger-live-mobile/src/mvvm/features/Q2WalletV4Tour/Drawer/hooks/useQ2WalletV4TourDrawerViewModel.ts
@@ -7,15 +7,7 @@ import { track, screen } from "~/analytics";
 import { PAGE_TRACKING_Q2_WALLET_V4_TOUR } from "../const";
 import type { Q2WalletV4TourDrawerViewModel } from "../types";
 
-export interface UseQ2WalletV4TourDrawerViewModelOptions {
-  /** When true and tour is enabled and not yet seen, the drawer will auto-open (e.g. on Portfolio page). */
-  readonly isOnPortfolioPage?: boolean;
-}
-
-export const useQ2WalletV4TourDrawerViewModel = (
-  options: UseQ2WalletV4TourDrawerViewModelOptions = {},
-): Q2WalletV4TourDrawerViewModel => {
-  const { isOnPortfolioPage = false } = options;
+export const useQ2WalletV4TourDrawerViewModel = (): Q2WalletV4TourDrawerViewModel => {
   const currentIndexRef = useRef(0);
   const isClosingRef = useRef(false);
   const hasSeenQ2WalletV4Tour = useSelector(hasSeenQ2WalletV4TourSelector);
@@ -29,14 +21,14 @@ export const useQ2WalletV4TourDrawerViewModel = (
     dispatch(setHasSeenQ2WalletV4Tour(true));
   }, [dispatch]);
 
-  const openDrawerWithTracking = useCallback((source: "portfolio" | "debug") => {
+  const openDrawerWithTracking = useCallback(() => {
     isClosingRef.current = false;
     currentIndexRef.current = 0;
     setIsDrawerOpen(true);
     screen(
       PAGE_TRACKING_Q2_WALLET_V4_TOUR,
       undefined,
-      { source: source === "portfolio" ? "Portfolio" : "Debug" },
+      { source: "Debug" },
       true,
       false,
     );
@@ -48,7 +40,7 @@ export const useQ2WalletV4TourDrawerViewModel = (
 
   const openQ2WalletV4Tour = useCallback(() => {
     if (!hasSeenQ2WalletV4Tour && isQ2TourEnabled) {
-      openDrawerWithTracking("debug");
+      openDrawerWithTracking();
     }
   }, [hasSeenQ2WalletV4Tour, isQ2TourEnabled, openDrawerWithTracking]);
 
@@ -75,12 +67,6 @@ export const useQ2WalletV4TourDrawerViewModel = (
     isClosingRef.current = true;
     handleCloseDialog();
   }, [handleCloseDialog]);
-
-  useEffect(() => {
-    if (isOnPortfolioPage && isQ2TourEnabled && !hasSeenQ2WalletV4Tour) {
-      openDrawerWithTracking("portfolio");
-    }
-  }, [isOnPortfolioPage, isQ2TourEnabled, hasSeenQ2WalletV4Tour, openDrawerWithTracking]);
 
   useEffect(() => {
     if (isDrawerOpen && !isQ2TourEnabled) {
