+++ apps/ledger-live-mobile/src/mvvm/features/Q2WalletV4Tour/Drawer/hooks/useSlideFooterButtonViewModel.ts
@@ -1,14 +1,22 @@
 import { useCallback } from "react";
 import { useSlidesContext } from "@ledgerhq/native-ui";
+import { interpolate, useAnimatedStyle } from "react-native-reanimated";
+import { useTranslation } from "~/context/Locale";
 import { track } from "~/analytics";
 import { PAGE_TRACKING_Q2_WALLET_V4_TOUR } from "../const";
 
 export const useSlideFooterButtonViewModel = (onComplete: () => void) => {
+  const { t } = useTranslation();
   const { totalSlides, currentIndex, goToNext, scrollProgressSharedValue } = useSlidesContext();
 
   const lastIndex = totalSlides - 1;
   const isFirstSlide = currentIndex <= 0;
   const isLastSlide = currentIndex >= lastIndex;
+  const fadeStart = lastIndex - 0.5;
+  const isInTest = process.env.NODE_ENV === "test";
+
+  const primaryLabel = isFirstSlide ? t("q2WalletV4Tour.cta.start") : t("q2WalletV4Tour.cta.next");
+  const doneLabel = t("q2WalletV4Tour.cta.done");
 
   const goNext = useCallback(() => {
     goToNext();
@@ -22,16 +30,42 @@ export const useSlideFooterButtonViewModel = (onComplete: () => void) => {
   const complete = useCallback(() => {
     onComplete();
     track("button_clicked", {
-      button: "Discover my new portfolio",
+      button: "Got it",
       page: PAGE_TRACKING_Q2_WALLET_V4_TOUR,
     });
   }, [onComplete]);
 
+  const continueStyle = useAnimatedStyle(
+    () => ({
+      opacity: interpolate(
+        scrollProgressSharedValue.value,
+        [fadeStart, lastIndex],
+        [1, 0],
+        "clamp",
+      ),
+    }),
+    [fadeStart, lastIndex, scrollProgressSharedValue],
+  );
+
+  const doneStyle = useAnimatedStyle(
+    () => ({
+      opacity: interpolate(
+        scrollProgressSharedValue.value,
+        [fadeStart, lastIndex],
+        [0, 1],
+        "clamp",
+      ),
+    }),
+    [fadeStart, lastIndex, scrollProgressSharedValue],
+  );
+
   return {
-    lastIndex,
-    isFirstSlide,
+    primaryLabel,
+    doneLabel,
+    continueStyle,
+    doneStyle,
     isLastSlide,
-    scrollProgressSharedValue,
+    isDoneButtonInteractive: isLastSlide || isInTest,
     goNext,
     complete,
   };
