+++ apps/ledger-live-mobile/src/mvvm/features/Q2WalletV4Tour/Drawer/types.ts
@@ -1,5 +1,3 @@
-import type { Q2WalletV4TourImageSource } from "./const";
-
 export interface Q2WalletV4TourDrawerViewModel {
   readonly isDrawerOpen: boolean;
   readonly openQ2WalletV4Tour: () => void;
@@ -9,9 +7,3 @@ export interface Q2WalletV4TourDrawerViewModel {
 }
 
 export type Q2WalletV4TourControls = Q2WalletV4TourDrawerViewModel;
-
-export type Q2WalletV4TourResolvedSlide = {
-  readonly title: string;
-  readonly subtitle: string;
-  readonly imageSrc: Q2WalletV4TourImageSource;
-};
